Remove deprecated Server, rename ClientSocket to Client
